This is a temporary commit
I am having issues with creating anchor links. When I create a link to
a heading or footnote within a document the anchor link is being
resolved relative to root rather than relative to the document
itself. In an attempt to resolve the issue I have published my test
page live because I wanted to find out if the issue was only happening
when rendering the site locally. It appears that it is an issue on the
published site as well. Once I have resolved the issue I will be
remooving my test page currently at https://bloggerbust.ca/post/anchor-links-not-working-in-hugo/