Switching to ox-hugo emacs package
I have been having difficulty getting hugo to render anchors local to
the document that contains them. Whenver I create a heading link or an
anchor in a document it is always being rendered local to the website
root directory. I have now switched to ox-hug because it is supposed
to have better support for org-mode. Unfortunately this has still not
resolved my issue, but I am going to make the switch anyway since it
is supposed to have better feature support for org-mode.