Fixed formatting issues and some typos
I noticed that the org formatting in this file was a mess for some
reason, so I cleaned it up. I also noticed that the source blocks do
not wrap. It is probably an issue with my theme. I removed whitespace
from some of the columns in the results output of some of the
commands, but it was not enough to fit everything in the width.