Created a tinyurl to the nauci base init licensed
Blackfriday defect
278 (https://github.com/russross/blackfriday/issues/278) causes
italicized links to render with underscores visible and the link not
italicized. To work around this defect the ox-hugo exporter that I use
to export org-mode markup to blackfriday markdown URI encodes
underscores in links. This should not be a problem except that
underscore percent-encoded characters are part of the unreserved set
which do not have standardized behaviour. It turns out that Github does
not treat URI encoded underscores the same as literal underscores and
so the workaround for the blackfriday defect exposed an issue when
linking to such URIs at Github. To get around this final issue I have
created a tinyurl for my license since the actual URI contains
underscores.

For reference see https://github.com/kaushalmodi/ox-hugo/issues/236.